actualizacion de delete y update tanto en usuario como en oferta de alquiler

+++ App/Controller/OfertaAlquilerController.php
@@ -112,10 +112,36 @@
 
         public function delete(){
             $result = new Result();
-    
-            $this->ofertaModel->deleteById($_POST['id']);
-            $result->success = true;
-            $result->message = "Oferta eliminada con éxito";
+
+            if ($_SERVER["REQUEST_METHOD"] == "POST") {
+                $idOferta = (isset($_POST['id']))?$_POST['id']:'';
+                if($idOferta != null && is_numeric($idOferta)){
+                    $oferta = $this->ofertaModel->getById($idOferta);
+                    if($oferta){
+                        $imagenes = explode(", ", $oferta['galeriaFotos']);
+                        foreach($imagenes as $imagen){
+                            $imagenPath = URL_PATH . "/Assets/images/fotoPerfil/" . $imagen;
+                            if (file_exists($imagenPath)) {
+                                unlink($imagenPath);
+                            }
+                        }
+                        $this->ofertaModel->deleteById($_POST['id']);
+                        $result->success = true;
+                        $result->message = "Oferta eliminada con éxito";
+                    }else{
+                        $result->success = false;
+                        $result->message = "Oferta no encontrada";
+                    }
+
+                }else{
+                    $result->success = false;
+                    $result->message = "Identificador invalido";
+                }
+            }else {
+                $result->success = false;
+                $result->message = "Solicitud no válida";
+            }    
+            
 
             echo json_encode($result);
         }
@@ -134,7 +160,92 @@
         }
 
         public function update(){
+            $result = new Result();
+            
+            if($_SERVER["REQUEST_METHOD"] == "POST"){
+                $idOferta = (isset($_POST['textID'])) ? $_POST['textID'] : null;
+                $titulo = (isset($_POST['titulo'])) ? $_POST['titulo']: '';
+                $descripcion = (isset($_POST['descripcion'])) ? $_POST['descripcion']: '';
+                $ubicacion = (isset($_POST['ubicacion'])) ? $_POST['ubicacion']: '';
+                $etiqueta = (isset($_POST['etiqueta'])) ? $_POST['etiqueta']: '';
+                $costo = (isset($_POST['costoAlquilerPorDia'])) ? $_POST['costoAlquilerPorDia']: '';
+                $cupo = (isset($_POST['cupo']))? $_POST['cupo']: '';
+                $tiempoMin = (isset($_POST['tiempoMinPermanencia'])) ? $_POST['tiempoMinPermanencia']: '';
+                $tiempoMax = (isset($_POST['tiempoMaxPermanencia'])) ? $_POST['tiempoMaxPermanencia']: '';
+                $fechaIni = (isset($_POST['fechaInicio'])) ? $_POST['fechaInicio']: '';
+                $fechaFin = (isset($_POST['fechaFin'])) ? $_POST['fechaFin']: '';
+                $listServicios = isset($_POST['listServicios']) ? implode(", ", $_POST['listServicios']) : '';
+                
 
+                
+
+                if($idOferta != null && is_numeric($idOferta)){
+                    $oferta = $this->ofertaModel->getById($idOferta);
+                    if($oferta){
+                        if (isset($_FILES['galeriaFotos']) && is_array($_FILES['galeriaFotos']['tmp_name'])) {
+                            // Procesar y subir nuevas imágenes
+                            $galeriaFotos = [];
+                            if (isset($oferta['galeriaFotos'])) {
+                                $imagenes = explode(", ", $oferta['galeriaFotos']);
+                                foreach ($imagenes as $imagen) {
+                                    $imagenPath = URL_PATH . "/Assets/images/fotoPerfil/" . $imagen;
+                                    if (file_exists($imagenPath)) {
+                                        unlink($imagenPath);
+                                    }
+                                }
+                            }
+                        
+                            foreach ($_FILES['galeriaFotos']['tmp_name'] as $key => $tmp_name) {
+                                // Validar si es una imagen y obtener la extensión del archivo
+                                $tipoMIME = $_FILES['galeriaFotos']['type'][$key];
+                                $extension = pathinfo($_FILES['galeriaFotos']['name'][$key], PATHINFO_EXTENSION);
+                        
+                                // Verificar que es una imagen válida
+                                if (strpos($tipoMIME, 'image') === 0) {
+                                    // Generar un nombre de archivo único con una extensión segura
+                                    $fecha_img = new DateTime();
+                                    $nombre_foto = $fecha_img->getTimestamp() . '_' . uniqid() . '.' . $extension;
+                        
+                                    // Ruta de destino relativa al directorio de imágenes
+                                    $destinationPath = 'Assets/images/galeriaFotos/' . $nombre_foto;
+                        
+                                    // Mover la imagen al directorio de destino
+                                    if (move_uploaded_file($tmp_name, $destinationPath)) {
+                                        $galeriaFotos[] = $nombre_foto;
+                                    }
+                                }
+                            }
+                            $fotosString = implode(", ", $galeriaFotos);
+                        } else {
+                            $fotosString = $oferta['galeriaFotos']; // No se subieron nuevas imágenes, mantener las antiguas
+                        }
+                        
+                        // Luego, actualiza la oferta con los datos proporcionados
+                        $this->ofertaModel->updateById($idOferta, [
+                            'titulo' => $titulo,
+                            'descripcion' => $descripcion,
+                            'ubicacion' => $ubicacion,
+                            'etiquetas' => $etiqueta,
+                            'galeriaFotos' => $fotosString, // Utilizar las nuevas imágenes o las antiguas
+                            'listServicios' => $listServicios,
+                            'costoAlquilerPorDia' => $costo,
+                            'tiempoMinPermanencia' => $tiempoMin,
+                            'tiempoMaxPermanencia' => $tiempoMax,
+                            'cupo' => $cupo,
+                            'fechaInicio' => $fechaIni,
+                            'fechaFin' => $fechaFin,
+                            'creadorID' => $id
+                        ]);
+                        
+                        $result->success = true;
+                        $result->message = "Oferta modificada con éxito";
+                        
+
+                }else{
+                    $result->success = false;
+                    $result->message = "Identificador invalido";
+                }
+            }    
         }
         
     }
